Move participant presenter to participants app
