+++ manage_breast_screening/record_a_mammogram/presenters.py
@@ -3,8 +3,8 @@ import re
 from django.urls import reverse
 
 from ..clinics.models import Appointment
+from ..participants.presenters import ParticipantPresenter
 from ..utils.date_formatting import format_date, format_relative_date, format_time
-from ..utils.string_formatting import format_age, format_nhs_number, sentence_case
 
 Status = Appointment.Status
 
@@ -100,32 +100,3 @@ class ClinicSlotPresenter:
         clinic = self._clinic
 
         return f"{format_time(clinic_slot.starts_at)} ({ clinic_slot.duration_in_minutes } minutes) - { format_date(clinic.starts_at) } ({ format_relative_date(clinic.starts_at) })"
-
-
-class ParticipantPresenter:
-    def __init__(self, participant):
-        self._participant = participant
-
-        self.extra_needs = participant.extra_needs
-        self.ethnic_group = participant.ethnic_group
-        self.full_name = participant.full_name
-        self.nhs_number = format_nhs_number(participant.nhs_number)
-        self.date_of_birth = format_date(participant.date_of_birth)
-        self.age = format_age(participant.age())
-        self.risk_level = sentence_case(participant.risk_level)
-
-    @property
-    def ethnic_group_category(self):
-        category = self._participant.ethnic_group_category()
-        if category:
-            return category.replace("Any other", "any other")
-        else:
-            return None
-
-    @property
-    def address(self):
-        address = self._participant.address
-        if not address:
-            return {}
-
-        return {"lines": address.lines, "postcode": address.postcode}
